write simple sample test code
